requirements.txt
# Core dependencies (required)
requests>=2.31.0
beautifulsoup4>=4.12.0
trafilatura>=1.6.0
pandas>=2.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
lxml>=4.9.0
lxml_html_clean>=0.1.0
psycopg2-binary
# AI providers (optional - chọn 1 hoặc nhiều)
openai>=1.0.0
google-generativeai>=0.3.0
autogen
anthropic

fastapi 
fastapi[standard]
alembic
qdrant_client
google-genai